server/requirements.txt
asgiref==3.7.2
coverage==7.3.2
Django==4.2.7
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0
pytz==2023.3.post1
setuptools==69.0.0
sqlparse==0.4.4
tzdata==2023.3
